+++ src/getJapaneseCalendarFormats.ts
@@ -1,0 +1,9 @@
+/**
+ * Returns a list of Japanese calendar formats.
+ * @return {string[]} An array of Japanese calendar format strings.
+ */
+export default (): string[] => [
+  'NNNNyoM月D日', 
+  'NNNNyoM月', 
+  'NNNNyo',
+];