+++ types/getJapaneseCalendarFormats.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns a list of Japanese calendar formats.
+ * @return {string[]} An array of Japanese calendar format strings.
+ */
+declare const _default: () => string[];
+export default _default;