+++ types/index.d.ts
@@ -1,4 +1,4 @@
-import jp2west from '~/jp2west';
-import west2jp from '~/west2jp';
-import isJapaneseCalendar from '~/isJapaneseCalendar';
-export { jp2west, west2jp, isJapaneseCalendar };
+import toWesternCalendar from '~/toWesternCalendar';
+import toJapaneseCalendar from '~/toJapaneseCalendar';
+import isValidJapaneseCalendar from '~/isValidJapaneseCalendar';
+export { toWesternCalendar, toJapaneseCalendar, isValidJapaneseCalendar };