+++ types/isValidJapaneseCalendar.d.ts
@@ -1,0 +1,8 @@
+import 'moment/locale/ja';
+/**
+ * Checks if the given value is a valid Japanese calendar date.
+ * @param {string} value The value to be validated.
+ * @return {boolean} if the value is a valid Japanese calendar date, `false` otherwise.
+ */
+declare const _default: (value: string) => boolean;
+export default _default;