+++ types/japaneseCalendarFormats.d.ts
@@ -1,6 +1,0 @@
-declare const _default: () => string[];
-/**
- * Get Japanese calendar format list.
- * @return {string[]} Japanese calendar format list.
- */
-export default _default;