+++ types/west2jp.d.ts
@@ -1,8 +1,0 @@
-import 'moment/locale/ja';
-declare const _default: (input: string) => string;
-/**
- * Convert a Western calendar date to a Japanese calendar date.
- * @param {string} input Western calendar date.
- * @return {string} Converts the Western calendar date to the Japanese calendar date and returns it.
- */
-export default _default;