+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2025/1/30
+### Changed
+- Updated moment from version 2.29.1 to 2.30.1
+- Updated rollup from version 2.67.3 to 4.32.1
+- Refactored function names:
+    - `west2jp` to `toJapaneseCalendar`
+    - `jp2west` to `toWesternCalendar`
+    - `isJapaneseCalendar` to `isValidJapaneseCalendar`
+    - `japaneseCalendarFormats` to `getJapaneseCalendarFormats`
+  - 
 ## [1.0.1] - 2023/11/7
 ### Added
 - Added Japanese calendar validation function (<code>isJapaneseCalendar()</code>).
@@ -10,3 +20,4 @@ All notable changes to this project will be documented in this file.
 - Released.
 
 [1.0.1]: https://github.com/takuya-motoshima/date-japanese/compare/v1.0.0...v1.0.1
+[2.0.0]: https://github.com/takuya-motoshima/date-japanese/compare/v1.0.1...v2.0.0